refactor(engine): peel pure helpers out of executor.ts
Move browser probe, requeue-loop signatures, pseudo-pause/handoff
detectors, and clearTerminalWorkflowStepFailures into executor/*
satellites with re-exports from executor.ts for stable test imports.
