+++ scripts/line-count-baseline.json
@@ -90,7 +90,7 @@
   "packages/engine/src/__tests__/triage.test.ts": 6605,
   "packages/engine/src/agent-heartbeat.ts": 5495,
   "packages/engine/src/agent-tools.ts": 4793,
-  "packages/engine/src/executor.ts": 19328,
+  "packages/engine/src/executor.ts": 19107,
   "packages/engine/src/merger-ai.ts": 2148,
   "packages/engine/src/merger.ts": 12487,
   "packages/engine/src/pi.ts": 2799,
